refactor(conf): move the aggregate-settings helpers to neutral ground
pool.py (supervisor side) refreshed the settings aggregate through
aleph.vm.agent.utils, the second-to-last supervisor -> agent import. The
settings-aggregate machinery is generic (fetch the aggregate from the API,
cache it, expose typed accessors) and is needed by both sides: the agent for
community-wallet/allocation-signer lookups, the supervisor for GPU
compatibility (pool.setup + resources.get_gpu_devices).

Move it to neutral modules under aleph.vm.utils:
- aleph.vm.utils.cache: AsyncTTLCache (moved from agent/cache.py)
- aleph.vm.utils.http: shared RetrySession/get_session with 429 retry
  (moved from agent/http.py)
- aleph.vm.utils.aggregate: AggregateSettingsDict, the settings cache,
  fetch/get/update_aggregate_settings and get_compatible_gpus join the
  existing user-aggregate helpers

agent/utils.py keeps the agent-only policy helpers (community wallet,
allocation signers, format_cost) on top of the neutral core. All importers
updated in place, no re-export shims; test patch targets follow the moves.
This also removes the indirect resources.py -> agent.utils edge the linter
could not see (allow_indirect_imports).
